reorganize and split up the source files
Instead of having all the code in the top level index file, split out
read and write into separate files and add a simple shared module to
stash the config.
